% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistral_preview_page.R
\name{mistral_preview_page}
\alias{mistral_preview_page}
\title{Preview Mistral OCR Page as HTML}
\usage{
mistral_preview_page(mistral_obj, page_num = 1)
}
\arguments{
\item{mistral_obj}{A Mistral OCR object containing pages with markdown content.}

\item{page_num}{The page number to preview (default: 1).}
}
\value{
A browsable HTML widget displaying the rendered page content.
}
\description{
This function converts the markdown content from a Mistral OCR object page
to HTML and displays it in a browsable format.
}
\examples{
\dontrun{
# Assuming test_mistral is a Mistral OCR object
mistral_preview_page(test_mistral, 1)
}

}
\author{
Nathan C. Layman
}
