% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_parse_response.R
\name{textract_extract_key_value_pairs}
\alias{textract_extract_key_value_pairs}
\title{Extract Key-Value Pairs from Textract Blocks}
\usage{
textract_extract_key_value_pairs(blocks)
}
\arguments{
\item{blocks}{List of Textract blocks}
}
\value{
Data frame with key-value pairs or NULL
}
\description{
Extract Key-Value Pairs from Textract Blocks
}
\keyword{internal}
