% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_parse_response.R
\name{textract_extract_tables}
\alias{textract_extract_tables}
\title{Extract Tables from Textract Blocks}
\usage{
textract_extract_tables(blocks)
}
\arguments{
\item{blocks}{List of Textract blocks}
}
\value{
List of data frames (one per table) or NULL
}
\description{
Extract Tables from Textract Blocks
}
\keyword{internal}
