% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_parse_response.R
\name{textract_get_block_text}
\alias{textract_get_block_text}
\title{Get Text Content from a Textract Block}
\usage{
textract_get_block_text(block, blocks)
}
\arguments{
\item{block}{Textract block}

\item{blocks}{All blocks for reference}
}
\value{
Character string with block text
}
\description{
Get Text Content from a Textract Block
}
\keyword{internal}
