% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_parse_response.R
\name{textract_parse_table}
\alias{textract_parse_table}
\title{Parse a Single Table from Textract Blocks}
\usage{
textract_parse_table(table_block, blocks)
}
\arguments{
\item{table_block}{Textract TABLE block}

\item{blocks}{All blocks for reference}
}
\value{
Data frame representing the table
}
\description{
Parse a Single Table from Textract Blocks
}
\keyword{internal}
